Fix UI contrast, guest detail view, dashboard activity, and multiple bugs
- Darken text colors for WCAG AA compliance (textPrimary, textSecondary, headings, status colors)
- Admin guest detail now returns all guest-submitted data (travel, hotel, food, dress, activities)
- Frontend GuestDetail shows all fields with correct backend field mapping
- Dashboard recent activity populated from real data (RSVPs, travel, hotel, activities, media, portal access)
- Hotels: star_rating changed to float for half-star support
- Dress codes: color palette normalized with both field name formats
- Guest upload: results populated from API response, pagination param fixed
- Export route moved before /{guest_id} to prevent route conflict
- Vite chunk splitting for performance (1126KB → 162KB main bundle)
- Guest portal: HotelSection/DressCodeSection use correct backend field names

+++ backend/app/schemas/hotel_schemas.py
@@ -51,7 +51,7 @@ class SuggestedHotelCreate(BaseModel):
     phone: Optional[str] = Field(None, max_length=50)
     distance_from_venue: Optional[str] = Field(None, max_length=100)
     price_range: Optional[str] = Field(None, max_length=100)
-    star_rating: Optional[int] = Field(None, ge=1, le=5)
+    star_rating: Optional[float] = Field(None, ge=1, le=5)
     description: Optional[str] = None
     amenities: Optional[List[str]] = None
     image_urls: Optional[List[str]] = None
@@ -84,7 +84,7 @@ class SuggestedHotelUpdate(BaseModel):
     phone: Optional[str] = Field(None, max_length=50)
     distance_from_venue: Optional[str] = Field(None, max_length=100)
     price_range: Optional[str] = Field(None, max_length=100)
-    star_rating: Optional[int] = Field(None, ge=1, le=5)
+    star_rating: Optional[float] = Field(None, ge=1, le=5)
     description: Optional[str] = None
     amenities: Optional[List[str]] = None
     image_urls: Optional[List[str]] = None
@@ -119,7 +119,7 @@ class SuggestedHotelResponse(BaseSchema):
     phone: Optional[str] = None
     distance_from_venue: Optional[str] = None
     price_range: Optional[str] = None
-    star_rating: Optional[int] = None
+    star_rating: Optional[float] = None
     description: Optional[str] = None
     amenities: Optional[List[str]] = None
     image_urls: Optional[List[str]] = None
